auth/requirements.txt
django==5.1.2
djangorestframework==3.15.2
djangogrpcframework==0.2.1
grpcio==1.67.0
grpcio-tools==1.67.0
pre_commit==4.0.1
ipython==8.29.0
mypy==1.13.0
django-environ==0.11.2
psycopg2-binary==2.9.10
